shifted logic for route directions async call
